chore(jobsdb): split jobsdb code in smaller logically grouped files
